[feat]: searchCoordinate api add, services forder path fix
