--- src/services/types/addressTypes.ts
@@ -1,36 +1,0 @@
-// Common 타입 정의
-export type TAddressCommon = {
-  errorMessage: string;
-  countPerPage: string;
-  totalCount: string;
-  errorCode: string;
-  currentPage: string;
-};
-
-// Juso 타입 정의
-export type TJuso = {
-  detBdNmList: string;
-  engAddr: string;
-  rn: string;
-  emdNm: string;
-  zipNo: string;
-  roadAddrPart2: string;
-  emdNo: string;
-  sggNm: string;
-  jibunAddr: string;
-  siNm: string;
-  roadAddrPart1: string;
-  bdNm: string;
-  admCd: string;
-  udrtYn: string;
-  lnbrMnnm: string;
-  roadAddr: string;
-  lnbrSlno: string;
-  buldMnnm: string;
-  bdKdcd: string;
-  liNm: string;
-  rnMgtSn: string;
-  mtYn: string;
-  bdMgtSn: string;
-  buldSlno: string;
-};
